### 2024-11-09 #7443 @crvernon
:wave: @fschncvg - You mention the following previous publications:
Chiasmus Detection: https://aclanthology.org/2021.latechclfl-1.11/
Metaphor Detection: https://aclanthology.org/2022.mwe-1.11/
Additional Detectors so far in this software package:  Polysyndeton - Epiphora - Alliterative verse A tool for people in stylometry, literary scholars, computational linguists/NLP researchers.

Please clearly state how / if the current submission is different from the above.  Thanks.
