### 2024-11-12 #7443 @fschncvg
The previous publications are the base forms of the chiasmus and metaphor detector. While both have code to replicate the experiments and could with some work probably used by someone proficient in programming to also detect those stylistic devices in other texts, this software provides a new implementation of the concept that can be either used as a python library in other programs or be used directly by e.g. literary scholars with no programming knowledge. The additional stylistic devices that can be detected (polysyndeton, epiphora, alliterative verse) do not rely on machine learning and have not been previously published by me, but since this software provides a growing general stylistic device detector collection, it makes sense to include them, especially since they are interesting for various stylometric analysis tasks.

So the main difference is: Two of the previously published papers provide methods that are implemented in this software package. Additionally this package provides three more methods to find other stylistic devices.
This software package provides a command line interface and a library for the(currently 5) methods to enable their use by both people proficient in python and researchers with linguistic, stylometric and literary expertise without programming knowledge.

About the similarity to those previously published papers the bot linked: My software package detect stylistic devices. Those linked do not.

@dpalsman published a software that analyzes other aspects of text (e.g. discourse markers, emotions...) and could be used together with my software to gain information the other does not address. We both use spaCy for preprocessing, which makes the interoperability easy. However, my package also provides a cltk backend for Middle High German which can be easily be extended for other classical languages supported by cltk.

@PetrKorab published a software that analyzes various parts of time-series structured text (e.g. news artices, social media posts) for various things, but no stylistic devices. However, my software could be integrated into theirs to also support stylistic devices.

@KennethEnvoldsen provides a software that augments text by e.g. standardising spelling and grammar. It does not compute stylometric information like the choice of words - or of stylistic devices. This could _maybe_ be used as a preprocessing step for my software, however it may destroy the stylistic devices that my software searches for.

@lmullen provides an R package for various tokenization tasks, but no stylistic device detection. It could be used as a preprocessing step, but since it is written in R, the interoperability to my python package is limited.

@jboynyc uses network analysis on collections of text to find out which words are used by different authors and how this connects and groups the authors. While it would be interesting to also do network analysis to find out how different authors use stylistic devices and how they are connected and grouped by that, their software does not use or find stylistic devices and provides a different service than mine.
