### 2024-11-25 #7443 @crvernon
:wave: @logological - are you available to edit this one?  I am going to pull it from scope query.
