### 2024-11-26 #7443 @sara-02
@logological I will be happy to help in reviewing this package, I am availbel to help post mid dec.
I checked the package and it seems that it currently does not have a sample example in the Readme or a link to documentation to follow @fschncvg what should be the starting point after one installs the library?
