### 2024-12-06 #7443 @logological
@mikekestemont @SamHames @rlskoeser Would you be willing and able to review this submission on stylistic device detection for the Journal of Open Source Software?
