### 2024-12-15 #7443 @logological
@beinborn @steffeneger Would you be willing and able to review this submission on stylistic device detection for the Journal of Open Source Software?
