### 2024-12-17 #7596 @editorialbot
[REVIEW]: FreeStylo: An easy-to-use stylistic device detection tool for stylometry
**Submitting author:** @fschncvg (<a href="http://orcid.org/0009-0008-9953-6695">Felix Schneider</a>)
**Repository:** https://github.com/cvjena/freestylo
**Branch with paper.md** (empty if default branch):
**Version:** 0.9.0
