### 2025-01-24 #7596 @logological
@sara-02, @rlskoeser How are you coming along with your reviews?  I haven't seen any activity in your checklists in the past month or so—if there are any issues you've identified that need action from the authors, please feel free to post a comment here or raise a separate issue in the paper's own GitHub repository.
