### 2025-01-31 cvjena/freestylo#1 @rlskoeser
Installation problems
Feedback on installation as part of JOSS review https://github.com/openjournals/joss-reviews/issues/7596

I initially tried to install with python 3.13 and got an error; when I switched to python 3.12 the package appeared to install properly. The minimum version is documented in package metadata, but it would be helpful to make the supported versions easier to find, perhaps in the readme.

The package is installed, but the command-line script is not evident. The current JOSS draft paper indicates that the command line tool should be called `stylotool`; the project setup indicates it should be called `freestylo` but neither is found in my virtualenv search path or bin directory. I would guess the script entry point is not configured correctly, but I'm not familiar with the syntax you're using.  You might consider updating to the more current pyproject.toml format - see https://learn.scientific-python.org/development/guides/packaging-simple/ for some nice documentation on options.

The python usage in the JOSS draft paper also seems to be out of date; when I replace `stylotool` with `freestylo` in the example import statements, then the imports suceeed.
