### 2025-01-31 cvjena/freestylo#2 @rlskoeser
Add example usage to readme
Documentation review for https://github.com/openjournals/joss-reviews/issues/7596

The JOSS software draft paper includes usage instructions for the library as both a command-line tool and as a python library, but the readme in the repository only includes installation instructions. The usage instructions should be made available directly from the repository, either in the readme or somewhere easily accessible from the readme.

I don't think this level of usage information is needed in the JOSS paper, so it may be that you can simply move the usage instructions from the JOSS paper into the readme.  Be sure to update to reflect the package rename from `stylotool` to `freestylo`.

The JOSS paper indicates that either spaCy or cltk can be used for preprocessing; it would be nice to provide examples of these.

It would also be helpful to provide a high-level list of supported stylistic devices in the readme and ensure that the documentation makes it possible for any of the supported modes to be used.
