### 2025-01-31 #7596 @rlskoeser
@fschncvg the software paper provides a very clearly discussion of the stylistic devices that are currently supported by the package, but it's missing is a higher level discussion. Why are these particular stylistic devices supported and not others? Is there some overlap between these particular devices that led to them being developed together? Were they created for particular research projects or particular interests of yours? I see that the chiasmus and metaphor detection is based on recent research; what about the others? Were any of them added because they were simply easy and thus "low-hanging fruit"? Are these the most common devices in German, or most common within some particular scope? You note the potential for expansion, is there anything on your roadmap that you are planning to add, or any devices you expect to be particularly easy or difficult to incorporate based on the architecture and approach? e.g., would sound-based literary devices like assonance or rhyme be any more or less difficult to detect than some that are already supported?  If there are research projects leveraging any of the stylistic detection , it would be great to include that in the paper as example applications of the software.

The other context that is missing in the software paper is a fuller discussion of the state of the field. I can well believe that there is a substantial gap between the broader, lower level text analysis and NLP software packages like spaCY, nltk, cltk, even book nlp and more specific literary tools, but surely this is not the only software package or tool in this space?  Are there other packages or tools, whether in python or other languages, perhaps R, that focus on particular literary devices? You should include this in the statement of need to more clearly situate your package and your contribution.

This is a smaller point, but you might want to more clearly differentiate your package and the focus on literary style from stylometry in general (and the R stylo package in particular), which is focused on a rather different aspect of literary style (particularly authorship attribution); the overlap in names may cause confusion otherwise.
