### 2025-03-11 cvjena/freestylo#2 @sara-02
After cloning the repo and running `bash example_chiasmus.sh` I see the following error
```
freestylo/src/stylotool.py': [Errno 2] No such file or directory
```
Consequently testing with the command mentioned in the paper itself
`stylotool --config config.json --input input.txt --output output.json` also gives the same error
```
stylotool: command not found
```

As @rlskoeser  mentioned, the name should be changed to freestylo.
However, when I replaced the standalone command with `freestylo --config config.json --input input.txt --output output.json` I still got the error
```
freestylo: command not found
```
