### 2025-03-11 #7596 @sara-02
@logological @fschncvg I went through the paper PDF as well as the project repo and faced the same issues that were reported. My current checklist reflects the issues below:

In terms of paper writing, my thoughts are aligned with @rlskoeser. The overall writing is good however, the content placement needs adjustment. Specifically to reiterate:

1. The usage section should be limited to a minimum, like stating the standalone tool. However, the rest of the information should be moved to the examples folder or the readme.
2. The current state of art comparison and statement of need section should also list some packages that do only one of the stylistic tasks. Right now, beyond the standard syntactic analysis by the likes of nltk and spacy, no other stylometric library exists in Python.

In terms of the repository, it is missing quite a few components and needs updation in other areas:
1. `pip install' is broken as it requires a specific Python version, which is not stated in the readme. @rlskoeser already reported the same, I have updated the issue with my experience with the tool.
2. Pytests failed two of the tests; I have created an issue with the same.
3. Docstring coverage is good for the src folder but missing in the test folder (would be good to have, low priority).
4. A minimum example to test the system is missing. The sh and command line examples are broken, and the library example requires an example.txt. A sample example.txt with essential components to test the repo should be provided in the readme. @rlskoeser already reported the same, I have updated the issue with my experience with the tool.
5. The contributor's guideline is missing, aka contributor.md, and it has missing info about setup, example, update and test workflow.
