### 2025-03-12 #7596 @fschncvg
Dear reviewers,

thank you for the comments.
I'll address your questions one-by-one:

@rlskoeser

Why are these particular stylistic devices supported and not others?
There was a need to start somewhere. In my research I developed methods for chiasmus detection and metaphor detection and especially for chiasmus detection there were no sufficient tools available. There was research on the topic, together with published code to replicate their experiments, but nothing that could directly used by us on new data.
So, especially metaphors and chiasmus are directly based on our research project (as part of [this project](https://dfg-spp-cls.github.io/projects_en/2020/01/24/TP-Anomaly-based_large-scale_analysis/))
The alliteration or Stabreim detection was included for a potential follow-up project, focusing also on old nordic literature that often uses this device.
Epiphora and polysyndeton detection were then added on the experiences we made when creating the detector for alliterative verse. Since they are stylistic devices that can be detected in a rule-based manner without machine learning models (at least after the preprocessing), similar to alliterative verse, we chose to add those.

Is there some overlap between these particular devices that led to them being developed together?
Alliterative verse, epiphora and polysyndeton overlap in their detection method. Their structure is relatively well-defined, and thus implementable in a rule-based form without relying on external training data.

Were they created for particular research projects or particular interests of yours?
As mentioned above, metaphor and chiasmus detection were developed as a part of this research project, alliterative verse detection was planned to be in the project and epiphora and polysyndeton were implementable in a similar way.

Were any of them added because they were simply easy and thus "low-hanging fruit"?
The "easiest" to implement was the alliteration detection, which was expanded to also find alliterative verse (longer alliterated sequences with some non-alliterative words inbetween) and which was included for a potential continuation of the project. This resulted in the idea of also implementing epiphora and polysyndeton detectors, which are also both rule-based, but considerably more complex in their structure.

Are these the most common devices in German, or most common within some particular scope?
Metaphors are very common in most languages, with e.g. the conceptual metaphor theory in cognitive linguistics finding metaphorical use of words in day-to-day language, like the phrase "paying attention".
The chiasmus is a device that was used at least in European languages since the antiquity, and chiastic structures also exists in e.g. texts in Semitic languages, like the Hebrew old testament. It was one goal of our research project to find out how the use of this device changed with both time and genre/topic in Middle High German (medieval) literature.
Alliterative verse is especially common in old Nordic texts, and "pure" alliterations still are used today.
Polysyndeton and Epiphora are both common in literature of various languages.

You note the potential for expansion, is there anything on your roadmap that you are planning to add, or any devices you expect to be particularly easy or difficult to incorporate based on the architecture and approach?
 e.g., would sound-based literary devices like assonance or rhyme be any more or less difficult to detect than some that are already supported?
Rhyme is a very special case as they would (probably) require more or different preprocessing or rely on word lists.
Assonance on the other hand could be implemented in a related way to alliterative verse. There was no plan to include that so far, but this looks like a promising idea.
For the addition of new devices there is no fixed road map. However, personally I would find it interesting to include more metaphor types and the related figure of simile (e.g. "she is a lion" vs "she fights like a lion"). Especially simile could be also implemented in a rule-based manner, especially using syntax trees.
Another device that should be added in later versions is related to the chiasmus, the parallelism. We found that our chiasmus detection also works well in detecting parallel A B A B patterns instead of A B B A patterns, however, more research is required for that to actually verify this approach.
Repetitions are also a candidate. A simple repetition detector would be easy to implement, but especially one that can also find repetitions with additional words in it would be more interesting ("e.g. how nice, how very nice").
Currently not planned from our side are device detection methods based on large models like BERT. The package is also geared towards literary scholars who often do not have the computational resources to run large models on large text collections.

The other context that is missing in the software paper is a fuller discussion of the state of the field. I can well believe that there is a substantial gap between the broader, lower level text analysis and NLP software packages like spaCY, nltk, cltk, even book nlp and more specific literary tools, but surely this is not the only software package or tool in this space? Are there other packages or tools, whether in python or other languages, perhaps R, that focus on particular literary devices? You should include this in the statement of need to more clearly situate your package and your contribution.
There were no programs that we found that were fitting for us - most of it was code from papers that was there to replicate results, but not to be easily runnable on new texts, and especially not collections of detectors. However, we will do another round of search and add related software to the paper.

This is a smaller point, but you might want to more clearly differentiate your package and the focus on literary style from stylometry in general (and the R stylo package in particular), which is focused on a rather different aspect of literary style (particularly authorship attribution); the overlap in names may cause confusion otherwise.
Do you suggest a change of the package name itself here? (freestylo vs stylo)
I can see the problem with "freestylo" sounding like a "free version of stylo" which is free already and does a different thing.

@sara-02

The usage section should be limited to a minimum, like stating the standalone tool. However, the rest of the information should be moved to the examples folder or the readme.
Ok, i will restructure that.

The current state of art comparison and statement of need section should also list some packages that do only one of the stylistic tasks. Right now, beyond the standard syntactic analysis by the likes of nltk and spacy, no other stylometric library exists in Python.
We will add more things to this. Should we also include mentions of "research code", meaning code that not a standalone tool, but to replicate paper findings?

`pip install' is broken as it requires a specific Python version, which is not stated in the readme. @rlskoeser already reported the same, I have updated the issue with my experience with the tool.
Pytests failed two of the tests; I have created an issue with the same.
Docstring coverage is good for the src folder but missing in the test folder (would be good to have, low priority).
A minimum example to test the system is missing. The sh and command line examples are broken, and the library example requires an example.txt. A sample example.txt with essential components to test the repo should be provided in the readme. @rlskoeser already reported the same, I have updated the issue with my experience with the tool.
The contributor's guideline is missing, aka contributor.md, and it has missing info about setup, example, update and test workflow.
We will look into these points and change things accordingly. When submitting the paper, we were asked to name a certain version of the package (I believe it was 0.5.0). Is it problematic for the submission if we bump up the version number for the changes, or can we do that?

Thank you both very much for your comments so far.
