### 2025-03-13 #7596 @sara-02
@fschncvg, thanks for your reply. From your discussion, there are no packages for understanding language style except for the standard NLP packages. Instead of citing all the paper codes, you can do the following:
1. Cite a survey paper that discusses the current research in this field.
2. You can reference closed source or enterprise tools that are doing similar work to show that a gap exists in open source tools for the same.
