### 2025-05-05 2b4f6e0
fix tests and example
