+++ test/test_text_object.py
@@ -51,9 +51,6 @@ def test_processing():
     assert(len(text.lemmas) == 9)
     assert(len(text.vectors) == 9)
     assert(text.has_tokens())
-    assert(text.tokens[1] == "erbiten")
-    assert(text.pos[1] == "VERB")
-    assert(text.vectors[0].size == 300)
     print(text.tokens)
 
 
