+++ paper.bib
@@ -80,3 +80,64 @@ year = {2022},
     pages = "214--217",
 }
 
+@misc{figurativechecker2023,
+  author       = {Izaz Ahmad},
+  title        = {Figurative Checker Online},
+  year         = {2023},
+  url          = {https://figurativechecker.com/},
+  note         = {Accessed: 2025-05-05}
+}
+
+@inproceedings{lai-etal-2023-multi,
+    title = "Multilingual Multi-Figurative Language Detection",
+    author = "Lai, Huiyuan and Toral, Antonio and Nissim, Malvina",
+    booktitle = "Findings of the Association for Computational Linguistics: ACL 2023",
+    month = July,
+    year = "2023",
+    address = "Toronto, Canada",
+    publisher = "Association for Computational Linguistics",
+}
+
+@inproceedings{kuhn2024antithesis,
+  title = {{Using Pre-trained Language Models in an End-to-End Pipeline for Antithesis Detection}},
+  booktitle = {Proceedings of the 2024 Joint International Conference On Computational Linguistics, Language Resources And Evaluation},
+  author = {K{\"u}hn, Ramona and Saadi, Khouloud and Mitrovi{\'c}, Jelena and Granitzer, Michael},
+  year = {2024},
+  month = may,
+  address = {Torino, Italy},
+}
+
+@article{Li2023MetaphorDV,
+  title={Metaphor Detection via Explicit Basic Meanings Modelling},
+  author={Yucheng Li and Shunyu Wang and Chenghua Lin and Guerin Frank},
+  journal={ArXiv},
+  year={2023},
+  volume={abs/2305.17268}
+}
+
+@misc{prowritingaid2025,
+  author       = {{ProWritingAid Team}},
+  title        = {ProWritingAid: The Storyteller's Toolkit},
+  year         = {2025},
+  url          = {https://prowritingaid.com/},
+  note         = {Accessed: 2025-05-05}
+}
+
+@misc{aussieai_literary_device_analyzer,
+  author       = {{AussieAI Team}},
+  title        = {Literary Device Analyzer},
+  year         = {2025},
+  url          = {https://www.aussieai.com/editor/literary-device-analyzer},
+  note         = {Accessed: 2025-05-05}
+}
+
+@inproceedings{dubremetz2017chiasmus,
+  author    = {Marie Dubremetz and Joakim Nivre},
+  title     = {Machine Learning for Rhetorical Figure Detection: More Chiasmus with Less Annotation},
+  booktitle = {Proceedings of the 21st Nordic Conference of Computational Linguistics (NODALIDA 2017)},
+  year      = {2017},
+  address   = {Gothenburg, Sweden},
+  publisher = {Link{\"o}ping University Electronic Press},
+  pages     = {146--155}
+}
+
