+++ paper.md
@@ -29,10 +29,10 @@ bibliography: paper.bib
 
 # Summary
 
-Stylistic devices are delibarately chosen linguistic expressions that 
-are used to convey a certain meaning or effect. They are often used in 
+Stylistic devices are delibarately chosen linguistic expressions that
+are used to convey a certain meaning or effect. They are often used in
 literature to create a certain atmosphere or to convey a certain message.
-Due to this matter, the detection of stylistic devices in text is an important 
+Due to this matter, the detection of stylistic devices in text is an important
 task in stylometry,the study of linguistic style. Often, finding these
 stylistic devices is a tedious and costly process that involves close reading
 of the texts, ideally by multiple experts. This is extra costly especially if
@@ -45,7 +45,7 @@ with a simple configuration file. It is designed to be usable by both experts
 and non-experts in programming. For those proficient in python, this package
 also provides a library with a collection of classes to detect stylistic
 devices in text, together with customizable text preprocessing (tokenizing,
-POS-tagging, etc.). 
+POS-tagging, etc.).
 
 The command-line tool is to be used on plain text files. It pre-processes the
 text, detects the stylistic devices specified in the configuration file, and
@@ -63,7 +63,19 @@ stylistic devices in text. While there exists a great variety of NLP libraries
 like nltk [@nltk], spaCy [@spacy], or cltk [@cltk] and command-line tools like
 CWB [@cwb], or UCS [@ucs] for the processing and low level analysis of text,
 there is a lack of tools that are specifically designed for the detection of
-stylistic devices. Information about the usage of stylistic devices is
+stylistic devices.
+Current options in this space would be for figurative language the online
+tool Figurative Checker [@figurativechecker] or the MMFLD framework [@lai-etal-2023-multi],
+which is a python framework. Other frameworks would be [@kuhn2024antithesis]
+for the detection of antithesis, [@Li2023MetaphorDV] for the detection of
+metaphors, [@schneider2021datadriven] for the detection of chiasmus,
+or [@dubremetz2017chiasmus] for antimetabole detection.
+
+However, most of those frameworks are not available as a ready-to-use tool, but as frameworks or code implementations of papers.
+A commercial tool that was designed as a help for writers is the ProWritingAid [@prowritingaid2025], which seems to be able to find various features of texts. However, the whole extend of this tool is not visible from their information material.
+Another commercial tool which is able to find stylistic devices such as animalification, similes, imagery, onomatopoeia, epizeuxis, and anadiplosis is the Literary Device Analyzer [@aussieai_literary_device_analyzer].
+
+Information about the usage of stylistic devices is
 important for many branches of stylometry, especially for the analysis of
 literary texts. This package aims to fill this gap by providing an easy-to-use
 tool and library for the detection of those stylistic devices. Due to its
@@ -75,7 +87,7 @@ other python programs, or as a stand-alone command-line tool.
 
 # Design and Supported Stylistic Devices
 
-The package contains a collection of approaches to detect stylistic devices 
+The package contains a collection of approaches to detect stylistic devices
 in text.
 By default, the preprocessing is done by spaCy[@spacy] or cltk[@cltk].
 The following stylistic devices are currently supported:
@@ -101,8 +113,8 @@ The chiasmus detector needs some special lists to function properly.
 
 - denylist: a list of Part of Speech (POS)-tags that are not allowed to be
   part of a chiasmus
-- allowlist: a list of POS-tags that are allowed to be part of a chiasmus. 
-Be careful: if such a list is given, all other POS-tags are not allowed to 
+- allowlist: a list of POS-tags that are allowed to be part of a chiasmus.
+Be careful: if such a list is given, all other POS-tags are not allowed to
[…]
