### 2025-05-05 #7596 @fschncvg
@logological I'm on it, thank you for the patience. The code should work now in version 0.6.0, the paper should follow in the coming days.
