### 2025-05-06 b212700
continuing paper and readme edits
