+++ README.md
@@ -4,7 +4,7 @@ An easy-to-use package for detecting stylistic devices in text. This package is 
 
 For those proficient in python, this package provides a collection of approaches to detect stylistic devices in text. For those less proficient in python, this package provides a simple interface to detect stylistic devices in text with simple commands and user-friendly configuration.
 
-# Usage example
+# Usage Example
 
 If after installation, run the following command in the root of the repository:
 
@@ -28,8 +28,111 @@ The package is available on PyPi and can be installed using pip.
 pip install freestylo
 ```
 
+# Usage Details
+
+The package can be used both as a library and as a stand-alone command-line tool.
+Both from the library and from the command-line tool, the results can be saved in a JSON file.
+This json file will contain the complete tokenized text.
+When using the functions from the library, the result will be a python container with a similar structure to the JSON file.
+
+## Standalone Tool
+
+The standalone version can be configured using a simple JSON configuration file. The file should specify the language of the text and the stylistic devices to detect. The following is an example configuration file:
+
+```json
+{
+    "language": "de",
+    "annotations": {
+        "chiasmus": {
+            "window_size": 30,
+            "allowlist": ["NOUN", "VERB", "ADJ", "ADV"],
+            "denylist": [],
+            "model": "/chiasmus_de.torch"
+        }
+    }
+}
+```
+
+You can then run the tool using the following command:
+
+```bash
+stylotool --config config.json --input input.txt --output output.json
+```
+
+This will read the text from the file `input.txt`, preprocess (tokenize, POS-tag, etc.) the text, detect the stylistic devices specified in the configuration file, and write the results to the file `output.json`.
+
+## Library
+
+The library comprises a collection of functions to detect the stylistic devices, as well as preprocessing based on spaCy.
+Should the user want to use different preprocessing or use the package with a different language than the supported ones, a TextObject can be created and filled with the needed manually computed contents.
+The stylistic device detectors can then be applied to the TextObject.
+
+A typical example would look like this:
+```python
+import numpy as np
+
+from stylotool import TextObject as to
+from stylotool import TextPreprocessor as tp
+from stylotool import ChiasmusAnnotation as ca
+from stylotool import MetaphorAnnotation as ma
+
+# first, create a TextObject from the raw text
+text = to.TextObject(
+        textfile = "example_textfile.txt",
+        language="en")
+
+# create a TextPreprocessor object and process the text
+# this does the tokenizing, lemmatizing, POS-tagging, etc.
+preprocessor = tp.TextPreprocessor(language="en")
+preprocessor.process_text(text)
+
+# you can also use a different preprocessing of your choice
+# without the TextPreprocessor object
+# just fill the TextObject with the needed contents
+# those could be provided e.g. by spaCy, nltk, cltk,
+# or any other method of your choice
+
+# many digital corpora are already tokenized and POS-tagged
+# they may come in various formats, such as TEI XML, CoNLL, etc.
+# if you have a text in those formats, you can fill the TextObject
+# with the needed contents
+# you can then fill the missing values in the TextObject
+# with e.g. word vectors or other features created with a method of your choice.
+
+# you can now add various annotations to the text object
+# here, we add a chiasmus annotation
+chiasmus = ca.ChiasmusAnnotation(
+        text=text)
+chiasmus.allowlist = ["NOUN", "VERB", "ADJ", "ADV"]
+chiasmus.find_candidates()
+chiasmus.load_classification_model("chiasmus_model.pkl")
+chiasmus.score_candidates()
+
+# here, we add a metaphor annotation
+metaphor = ma.MetaphorAnnotation(
+        text=text)
+metaphor.find_candidates()
+metaphor.load_model("metaphor_model.pkl")
+metaphor.score_candidates()
+
+# finally, save the annotated text to a json file
+text.serialize("annotated_text.json")
+
[…]
