+++ README.md
@@ -4,6 +4,14 @@ An easy-to-use package for detecting stylistic devices in text. This package is 
 
 For those proficient in python, this package provides a collection of approaches to detect stylistic devices in text. For those less proficient in python, this package provides a simple interface to detect stylistic devices in text with simple commands and user-friendly configuration.
 
+# Installation
+
+The minimum python requirement for this package is python 3.12.
+The package is available on PyPi and can be installed using pip.
+
+```
+pip install freestylo
+```
 # Usage Examples
 
 ## Standalone Tool
@@ -15,21 +23,13 @@ freestylo --input test/documents/chiasmustext.txt --output ./output.json --confi
 ```
 
 This creates the file `output.json` in the root of the repository, which contains the detected stylistic devices in the text file `test/documents/chiasmustext.txt`.
-Afterewards, run the following command to get an overview over the results:
+Afterwards, run the following command to get an overview over the results:
 
 ```bash
 freestylo --mode report --data output.json --device chiasmustext
 ```
 
-# Installation
-
-The minimum python requirement for this package is python 3.12.
-The package is available on PyPi and can be installed using pip.
-
-```
-pip install freestylo
-```
-
+The report mode is currently just implemented for Chiasmus.
 
 The package can be used both as a library and as a stand-alone command-line tool.
 Both from the library and from the command-line tool, the results can be saved in a JSON file.
