### 2025-05-07 1fb5f25
updated bibliography
