+++ paper.bib
@@ -141,3 +141,30 @@ year = {2022},
   pages     = {146--155}
 }
 
+@article{eder2016stylor,
+  author = {Maciej Eder and Jan Rybicki and Mike Kestemont},
+  title = {{Stylometry with R: A Package for Computational Text Analysis}},
+  year = {2016},
+  journal = {{The R Journal}},
+  doi = {10.32614/RJ-2016-007},
+  url = {https://doi.org/10.32614/RJ-2016-007},
+  pages = {107--121},
+  volume = {8},
+  number = {1}
+}
+
+@misc{coles2019rhyme,
+  author       = {A. Coles},
+  title        = {Deep Rhyme Detection},
+  year         = {2019},
+  url          = {https://github.com/a-coles/deep-rhyme-detection},
+  note         = {Accessed: 2025-05-07}
+}
+
+@misc{marozick2021rap,
+  author       = {Alex Marozick and Abduarraheem Elfandi and Juno Mayer},
+  title        = {RapAnalysis: Rhyme Detection and Analysis Applied to User's Spotify Data},
+  year         = {2021},
+  url          = {https://github.com/alexmarozick/RapAnalysis},
+  note         = {Accessed: 2025-05-07}
+}
