+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = freestylo
-version = 0.6.0
+version = 0.7.0
 description = A tool for stylistic device detection."
 author = Felix Schneider
 author_email = felix.schneider@uni-jena.de
