+++ .github/workflows/package-and-publish.yml
@@ -57,7 +57,7 @@ jobs:
         --user
     - name: Run the tests
       run: python3 -m pytest
-    
+
 
 
 
@@ -67,6 +67,7 @@ jobs:
     if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
     needs:
     - build
+    - run_tests
     runs-on: ubuntu-latest
     environment:
       name: pypi
