+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = freestylo
-version = 0.7.0
+version = 0.8.1
 description = A tool for stylistic device detection."
 author = Felix Schneider
 author_email = felix.schneider@uni-jena.de
