### 2025-05-08 cvjena/freestylo#1 @fschncvg
Thank you for the feedback. For now I  updated the readme reflecting the python 3.12 requirement.
More recent python versions will be supported in the future to improve the usability of the library ( #4 )
