+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = freestylo
-version = 0.8.1
+version = 0.8.2
 description = A tool for stylistic device detection."
 author = Felix Schneider
 author_email = felix.schneider@uni-jena.de
