### 2025-05-08 cvjena/freestylo#2 @fschncvg
Thank you for the feedback! The tool should work now as expected. Also I have moved the usage section from the paper to the readme and added another test as an example on how to work with external preprocessing.
