### 2025-06-01 cvjena/freestylo#6 @sara-02
Sample example in Readme is giving error
The example code provided in the [Library](https://github.com/cvjena/freestylo?tab=readme-ov-file#library) section is throwing an error at the line `chiasmus.load_classification_model("chiasmus_model.pkl")'

```
site-packages/freestylo/Configs.py", line 75, in get_model_path
    raise FileNotFoundError(f"Model {model_to_load} not found")
FileNotFoundError: Model /home/sarah/.freestylo/models/chiasmus_model.pkl not found

```
