### 2025-06-01 #7596 @sara-02
@logological @fschncvg I have revisited my reviews, thanks for all the changes and improvements, there are 2 set of issues that still need attention.
1. The pip install and pytest is working now, and 2 standalone examples are also working. However the example code in the `Library' paragraph of the README is not working. I have created issue  for the same.  https://github.com/cvjena/freestylo/issues/6
Do let me know if I am missing something in my install and if that needs to be mentioned in the readme as well.
3. There are a couple of writing changes in the new PDF:- Line 25 and 132 mention the package name as `stylotool' it should be `freestylo' instead.
- Line 30 reference is missing.
- Line 106 Instead of `Since for..', we can have `For ...'  instead.
