### 2025-06-02 cvjena/freestylo#6 @rlskoeser
I ran into this as well but got interrupted before I opened an issue.

The test document and example configuration aren't included in the python package - and it's not clear to me that they _ought_ to be (perhaps the sample configuration). The directions seem like they are written for someone working with a clone of the GitHub repo.   My suggestion is to update the documentation to make it easy to download or create these example files for anyone who wants to use them as a starting point.
