### 2025-06-02 cvjena/freestylo#7 @rlskoeser
Documentation doesn't help me understand how to find & configure supported stylistic devices and available models
The readme promises "an overview of the detectors and their methods" and links to the documentation in docs.md, but there is no overview  in the [documentation](https://github.com/cvjena/freestylo/blob/main/docs/docs.md). It looks like code-level API documentation - it doesn't give me the information I need to update the configuration (or write my own) for a specific style or models.

Please provide information needed to configure each supported style, including all available models.

The readme lists the supported stylistic devices but doesn't specify language; absent a list of models it's hard for me to tell, are these German language only? I recommend listing the supported languages parenthetically after the style (e.g. "Alliteration (German)") - unless they are all currently the same set of languages, in which case you could just mention it once prominently.

FYI the formatting in the documentation also isn't displaying the content very well on GitHub; the parameters are running together without line breaks.
