### 2025-06-02 cvjena/freestylo#8 @rlskoeser
Document where models are stored and if that path is configurable
I noticed when I ran the example that models were downloaded - I had to go hunting around to discover they were stored in a `.freestylo/models/` directory in my home directory. At a minimum you should document this; if it's possible to override it with an environment variable or config parameter, that would be helpful as well.
