### 2025-06-02 #7596 @rlskoeser
Thanks to @sara-02 for opening https://github.com/cvjena/freestylo/issues/6 - I've commented with my experience and notes on the same problem.

I also opened two issues related to documentation:
- https://github.com/cvjena/freestylo/issues/7
- https://github.com/cvjena/freestylo/issues/8

The first one is a blocker for me actually understanding the available annotations and models well enough to configure and run the tool on my own content.
