### 2025-10-17 #7596 @rlskoeser
Hi, all. @fschncvg your updates to the code and revisions to the paper look great.  The report output for chiasmus detection is very helpful, I was able to run some familiar english examples of some of these devices through, and I'm getting meaningful output.

I've opened a PR with some suggestions for a few small improvements to your readme based on following the instructions as written:
https://github.com/cvjena/freestylo/pull/9

I also opened two issues:
- https://github.com/cvjena/freestylo/issues/11 this is related to the example code in the readme, which I think should be addressed before I sign off on the example use.  (I'm assuming the example should run as written, perhaps I misunderstood?)
- https://github.com/cvjena/freestylo/issues/10 reporting an exception I ran into when I used english text but had it configured for german; that doesn't need to be resolved for me to sign off, but I thought you'd want to know.

It might be helpful to clarify the output format in your documentation a bit, it took me some looking to figure out that the id was the token index (the report was helpful to ensure I was understanding correctly). I see now that the fields are is listed in your docs, but maybe one short example of a stylistic device and corresponding json output would help.

I'm ready to sign off as soon as the readme instructions and example code are corrected.
