### 2025-12-07 #7596 @fschncvg
Hi, thank you for the patience. I fixed the errors and also expanded the README.md with information about the output .json file.
