### 2025-12-14 #7596 @fschncvg
Finally a version worked. I pulled the model download to an earlier step and the models work.
The version on PiPy is version 0.8.8.
