### 2025-12-29 #7596 @logological
@fschncvg I have also opened an issue cvjena/freestylo#17 concerning contradictory licensing statements in the repository.  Could I please ask you to fix this?  JOSS can't publish the paper unless the accompanying software has an unambiguously stated free or open source licence.  (I didn't create a pull request for this particular change as I think it's important that it be applied by the code's primary author.)
