### 2026-01-17 #7596 @fschncvg
Done!
Version number: 0.9.0
DOI: 10.5281/zenodo.18281153
[https://doi.org/10.5281/zenodo.18281152](https://doi.org/10.5281/zenodo.18281152)
