### 2026-01-19 #7596 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
