### 2026-01-23 #7596 @crvernon
@fschncvg - I just need you to make the following changes before I accept:

In your archive:

- [x] The name of your Zenodo archive must match the title of your paper exactly.  Please edit the metadata in your Zenodo archive to reflect this.  No new release required.

Everything else looks great.  Let me know when this change has been made.
