### 2026-01-26 #7596 @crvernon
@fschncvg they still do not match exactly.  See:

Title of the archive:  FreeStylo - an easy-to-use stylistic device detection tool for stylometry
Title of the paper:  FreeStylo: An easy-to-use stylistic device detection tool for stylometry

Please edit the archive to be exactly as the paper title appears.  Thank you.
