### 2026-01-27 #7596 @crvernon
🥳 Congratulations on your new publication @fschncvg ! Many thanks to @logological for editing and @sara-02 and @rlskoeserfor your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
